requirements.txt
# Elasticsearch 6.x
elasticsearch-dsl>=7.0.0,<8.0.0
piper
loopercli
bbconf>=0.0.1